fix: file and image route refactoring
